Sync notes PoC
This PoC make syncronization works.

And by syncronization I mean `git add -A && git commit -m '/note/path'`
Which starting to sound a lot awful actually. But I'll work on that
later.

I think a good improvement now is refactor this procedural code into
something more like an API, with a nice struct to hold a note and it's
information regarding if is editing or creating a new one.